+++ src/vue/pages/Signup.vue
@@ -90,6 +90,7 @@ export default {
       storeWallet: vuexTypes.STORE_WALLET,
       loadAccount: vuexTypes.LOAD_ACCOUNT,
       loadKyc: vuexTypes.LOAD_KYC,
+      loadKvEntriesAccountRoleIds: vuexTypes.LOAD_KV_ENTRIES_ACCOUNT_ROLE_IDS,
     }),
     handleChildFormSubmit (form) {
       this.email = form.email
@@ -124,6 +125,7 @@ export default {
             },
           })
         }
+        this.loadKvEntriesAccountRoleIds()
       } catch (e) {
         ErrorHandler.process(e)
       }