+++ envs/local.js
@@ -1,6 +1,6 @@
 module.exports = {
   NODE_ENV: '"default"',
-  HORIZON_SERVER: '"https://api.anubis.tokend.io"',
-  FILE_STORAGE: '"https://s3-eu-west-1.amazonaws.com/anubis-api"',
-  NETWORK_PASSPHRASE: '"anubis Network!"',
+  HORIZON_SERVER: '"https://api.angus.tokend.io"',
+  FILE_STORAGE: '"https://s3-eu-west-1.amazonaws.com/angus-is"',
+  NETWORK_PASSPHRASE: '"Angus Network!"',
 }