+++ src/vue/pages/Signup.vue
@@ -103,7 +103,7 @@ export default {
       this.disableForm()
       try {
         const { response, wallet } = await Sdk.api.wallets.create(
-          this.email,
+          this.email.toLowerCase(),
           this.password,
           this.recoveryKeypair
         )
@@ -119,7 +119,7 @@ export default {
             ...vueRoutes.verify,
             params: {
               paramsBase64: btoa(JSON.stringify({
-                email: wallet.email,
+                email: wallet.email.toLowerCase(),
                 walletId: wallet.id,
               })),
             },